Use default rather than grab cursor on map (fixes #341)
Although the grab cursor is nice because it indicates that the map is
draggable, on many maps we need to click with some precision, e.g. to
get meteorological information on the point we click, or to set the
location of an agrifield. The grab cursor's point isn't clear, whereas
in the default cursor it's very clear. Incidentally, Google Maps also
uses a default cursor although it's draggable.